// generated/backend/models/NotebookAiAssistant.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Notebook } from './Notebook';
export type NotebookAiAssistant = {
    id: number;
    notebook?: Notebook;
    additionalInstructionsToAi?: string;
    createdAt?: string;
    updatedAt?: string;
};